visual changes, added items are loading to all tables

+++ ForWaiters/Models/Dish.cs
@@ -74,7 +74,7 @@ namespace ForWaiters.Models
             this.priceForFew = this.priceForOne * this.quantity;
         }
 
-        public int getPriceForFew()
+        public float getPriceForFew()
         {
             return this.priceForOne * quantity;
         }
